+++ examples/scaling.rs
@@ -1,7 +1,6 @@
 use tetra::graphics::scaling::{ScalingMode, ScreenScaler};
 use tetra::graphics::text::{Font, Text};
-use tetra::graphics::ui::NineSlice;
-use tetra::graphics::{self, Color, Rectangle, Texture};
+use tetra::graphics::{self, Color, NineSlice, Rectangle, Texture};
 use tetra::input::{self, Key};
 use tetra::math::Vec2;
 use tetra::{Context, ContextBuilder, Event, State};
@@ -16,7 +15,8 @@ const PANEL_Y: f32 = (SCREEN_HEIGHT / 2.0) - (PANEL_HEIGHT / 2.0);
 
 struct GameState {
     scaler: ScreenScaler,
-    panel: NineSlice,
+    panel_texture: Texture,
+    panel_config: NineSlice,
     text: Text,
 }
 
@@ -24,12 +24,8 @@ impl GameState {
     fn new(ctx: &mut Context) -> tetra::Result<GameState> {
         Ok(GameState {
             scaler: ScreenScaler::with_window_size(ctx, 640, 480, ScalingMode::Fixed)?,
-            panel: NineSlice::new(
-                Texture::new(ctx, "./examples/resources/panel.png")?,
-                PANEL_WIDTH,
-                PANEL_HEIGHT,
-                Rectangle::new(4.0, 4.0, 24.0, 24.0),
-            ),
+            panel_texture: Texture::new(ctx, "./examples/resources/panel.png")?,
+            panel_config: NineSlice::with_border(Rectangle::new(0.0, 0.0, 32.0, 32.0), 4.0),
             text: Text::new(
                 format!("{}\n{:?}", LABEL, ScalingMode::Fixed),
                 Font::vector(ctx, "./examples/resources/DejaVuSansMono.ttf", 16.0)?,
@@ -65,7 +61,13 @@ impl State for GameState {
         graphics::set_canvas(ctx, self.scaler.canvas());
         graphics::clear(ctx, Color::rgb(0.392, 0.584, 0.929));
 
-        self.panel.draw(ctx, Vec2::new(PANEL_X, PANEL_Y));
+        self.panel_texture.draw_nine_slice(
+            ctx,
+            &self.panel_config,
+            PANEL_WIDTH,
+            PANEL_HEIGHT,
+            Vec2::new(PANEL_X, PANEL_Y),
+        );
         self.text.draw(ctx, Vec2::new(PANEL_X + 8.0, PANEL_Y + 8.0));
 
         graphics::reset_canvas(ctx);